feat(home): restructuring homepage
