+++ src/pages/Home/Map/styles.css
@@ -1,0 +1,8 @@
+.map {
+  height: 400px;
+  width: 400px;
+}
+
+.marker {
+  margin-top: -30px;
+}
