--- src/pages/Home/styles.css
@@ -1,8 +1,0 @@
-.map {
-    height: 400px;
-    width: 400px;
-  }
-  
-  .marker {
-    margin-top: -30px;
-  }
